[Testing hyperparameter tuning], seems working better. most significant one was learning rate. also dataset is important. in this commit, I used training_data/p4096_s1024_rps10_spp_20_ndp_100/rl/

+++ benchmarks/data-driven-routing/training/routing_agent/routing-agent-service/run_offline_training.sh
@@ -1,15 +1,27 @@
 #!/bin/bash
 
 
-training_data_dir="training_data/p4096_s1024_rps10_spp_20_ndp_100/prefix"
+# training_data_dir="training_data/p4096_s1024_rps10_spp_20_ndp_100/prefix"
+
+# training_data_dir="training_data/p4096_s1024_rps10_spp_20_ndp_100/random"
+
+# training_data_dir="training_data/p4096_s1024_rps20/prefix"
+# training_data_dir="training_data/p4096_s1024_rps20/random"
+training_data_dir="training_data/p4096_s1024_rps20/rl"
+# training_data_dir="training_data/p4096_s1024_rps20"
+
 final_model_dir="${training_data_dir}/final_model"
-rm -r ${final_model_dir}
-echo "Removed existing final model directory: ${final_model_dir}"
+if [ -d "${final_model_dir}" ]; then
+    echo "Final model directory already exists: ${final_model_dir}"
+    rm -r ${final_model_dir}
+    echo "Removed existing final model directory: ${final_model_dir}"
+fi
+mkdir -p "${final_model_dir}"
 
 # Define the command without backgrounding it initially
 python_cmd="python3 offline_routing_agent.py ${training_data_dir}/data.csv --model simpler_contextual_bandit --analyze_behavior --ttft_slo 1000 --avg_tpot_slo 50"
 
-# Execute the command, redirecting output
+## Execute the command, redirecting output
 ${python_cmd} &> output.log
 # ${python_cmd}
